Removed ESP8266 support
With the addition of the embedded webserver the ESP8266 ran into memory
issues when using newer versions of the ESP-Arduino framework. Since the
framework updates include security fixes it's not practical to keep it
pinned to version 2.6.3 in the long run. Dropping ESP8266 support seems
like the best option going forward.